LayoutTests/http/tests/xmlhttprequest/send-data-view-async-expected.txt

Test for array buffer and data view POST in XMLHttpRequest async send.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


req.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/post-echo-as-ascii.cgi
PASS status is "200"
PASS responseText is "0 1 2 48 49 50 128 129 130 253 254 255"
req.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/post-echo-as-ascii.cgi
PASS status is "200"
PASS responseText is "0 1 2 48 49 50 128 129 130 253 254 255"
PASS successfullyParsed is true

TEST COMPLETE
